lead-scorer/requirements.txt
psycopg2-binary==2.9.7
redis==4.6.0
requests==2.31.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3